Switch from rely/deny to ds-auth
Modules using rely/deny for auth must make use of a proxy or 'bridge'
contract to interface with modules that use the ds-auth scheme.

Since the rest of dappsys (including ds-chief which this module is
intended to be used with) uses ds-auth, a non trivial reduction in
deployment and auth complexity can be realised by replacing rely/deny
with ds-auth, and doing away with the bridge contracts.