Add value and gas to interface
This moves towards having all paramters of the call in exec specified as
part of plan.

Previously the caller of exec could choose which msg.value and msg.gas
the delegatecall used. A malicious attacker could have planed a call to
some innocuous looking code that had concealed gas / value dependent
behaviour, and then crafted the call to exec to trigger the hidden
behaviour.