Isolated storage for delegatecall
The code executed by delegatecall has complete access to the pause's
storage. This is messy and makes plans hard to audit as they can
potentially do some pretty sneaky stuff (e.g. bypass the delay by
writing directly to the plans mapping).

In this change we pull the code that does the actual delegatecall out
into its own contract (Delegator), making it impossible for the
delegatecall to directly modify the storage of the pause.

An additional check in exec is also required to make sure that the
storage of the delegator itself remains unmodified.

This change does probably make reasoning about the pause and its place
in some auth scheme a little harder as it now effectively has two
identities (DSPause & Delegator).