Remove eta from plot interface
Previously plotting plans from off chain (e.g. using seth) was pretty
clunky, as you didn't know what the value of block.timestamp would be
when the transaction was actually included in a block. This meant that
you had to guess and set eta to a value far enough in the future that
eta would still be after block.timestamp + delay when the tx was mined.

Here we fix this issue by removing the `eta` parameter from the
interface to plot. We always set eta to be now + delay in plot and then
plot returns the value of eta to be used in exec.