Revert "Use solidity delegatecall"
This reverts commit 4910b608559940c43d3011dbc06a8c608442b82a.

We don't know what bytecode solidity will generate for it's built in
delegatecall operation, making it impossible to guarantee that the gas
provided to the delegatecall will always match that specified in the
plan (solidity could always insert an arbitrarily large number of
instructions between any checks we do and the execution of the
delegatecall opcode itself).

We therefore revert to the assembly implementation from ds-proxy (we can
be sure that there are no instructions between our gas checks and the
execution of the delegatecall).