Specify gas as part of plan
Anyone in the world can execute a plan, and it seems a little reckless
to give them any control over the parameters of the delegatecall.

To provide a concrete example, an attacker could trick authorized
users into plotting an innocent looking plan containing malicious gas
dependent behaviour. The caller to exec would then provide the specific
amount of gas that would trigger the attack.

To prevent this, users must now provide the amount of gas that will be
passed to the delegatecall when they plot a plan.

I have done my best to encode the restrictions implied by EIP-150 as
accurately as possible, but as the calculations required themselves
consume gas, it is not possible to produce an implementation that is
both:

- accurate
- robust against future compiler / gas schedule changes

I have chosen to prioritise the second, and so the caller must provide a
little more gas than is strictly nescessary.

Two new invariants are enforced by this change

- The caller of exec cannot influence the parameters of the delegatecall
- A plan can only be executed if the caller provides sufficient gas