Specify extcodehash in plans
Since the introduction of the CREATE2 opcode we can no longer assume
that the same address will always contain the same code
([details](https://ethereum-magicians.org/t/potential-security-implications-of-create2-eip-1014/2614)).

Here, we defend against this attack by forcing callers of plot to
specify the expected codehash of usr. When exec is called, the codehash
is compared to the tag, and exec will revert if they do not match.